% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/condition.R
\name{condition}
\alias{condition}
\alias{error}
\alias{is.condition}
\alias{is.error}
\alias{is.tryError}
\alias{is.warning}
\title{Custom Conditions S3-Classes}
\usage{
condition(subclass = NULL, message, call = sys.call(-1))

error(subclass = NULL, message, call = sys.call(-1))

is.condition(x)

is.error(x)

is.tryError(x)

is.warning(x)
}
\arguments{
\item{subclass}{character, \emph{unique} name(s) subclass(es) that
extend the base classes \code{class}, see details in
\code{\link{extend_class}()}}

\item{message}{character, condition message}

\item{call}{call stack,}

\item{x}{an object to check for inheritance}
}
\value{
The function returns an S3 class of type \code{subclass} inheriting
 from \code{error} (only for \code{error()}) and \code{condition}.
}
\description{
The functions are wrapper for functions \code{\link[base]{simpleCondition}}
and \code{\link[base]{simpleError}} extending the classes with a custom
subclass. The resulting S3 class objects can be used to signal a custom
condition or, more specific, a custom error.

Functions \code{is.condition()}, \code{is.error()},
 \code{is.warning()}, and \code{is.tryError()} check for inheritance from
 abstract S3 class \code{condition}, \code{error}, \code{warning}, or
 \code{try-error} respectively. The latter can be used to test results of
 \code{\link[base]{try}()} to handle error-recovery.
}
\seealso{
\code{\link[base]{simpleCondition}} and
  \code{\link[base]{simpleError}} for general details on condition handling
}
