% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exceptionHandling.R
\name{exceptionHandling}
\alias{exceptionHandling}
\alias{custom_stop}
\alias{custom_warning}
\alias{custom_message}
\alias{custom_signalCondition}
\alias{custom_exceptionHandler}
\title{Customized Exception Handling}
\usage{
custom_stop(..., subclass = NULL, call. = TRUE, domain = NULL)

custom_warning(..., subclass = NULL, call. = TRUE, immediate. = FALSE,
  noBreaks. = FALSE, domain = NULL)

custom_message(..., subclass = NULL, domain = NULL, appendLF = TRUE)

custom_signalCondition(subclass = NULL, message, call = sys.call(-1),
  type = c("condition", "error", "warning", "message"))

custom_exceptionHandler(subclass = NULL, type = c("condition", "error",
  "warning", "message"))
}
\arguments{
\item{...}{zero or more objects which can be coerced to character or a
single condition, see \code{\link[base]{stop}()},
\code{\link[base]{warning}()}, or \code{\link[base]{message}()} for
details}

\item{subclass}{\emph{unique} name(s) of subclass(es) of the customized
conditions}

\item{call.}{passed to \code{\link[base]{stop}} or
\code{\link[base]{warning}}, for details see references therein}

\item{domain}{passed to \code{\link[base]{stop}},
\code{\link[base]{warning}} or \code{\link[base]{message}}, for details
see references therein}

\item{immediate.}{passed to \code{\link[base]{warning}}, for details see
references therein}

\item{noBreaks.}{passed to \code{\link[base]{warning}}, for details see
references therein}

\item{appendLF}{passed to \code{\link[base]{message}}, for details see
references therein}

\item{message}{character string}

\item{call}{call expression, see \code{\link[base]{simpleCondition}} for
details}

\item{type}{character string, type of the condition}
}
\value{
The function \code{exceptionHandler} returns a function with
  arguments \code{...} (see arguments above for a description). Customized
  exception handlers invoke the basic functions \code{\link[base]{stop}()},
  \code{\link[base]{warning}()}, or \code{\link[base]{message}()}. Note that
  the customized exception handler lacks any optional arguments of these
  functions.
}
\description{
The functions provide interfaces to raise customized errors, warnings, or
messages.
}
\details{
For signalling customized condtions using
  \code{custom_signalCondition} no exception handling is invoked even for
  types \code{"error"}, \code{"warning"}, or \code{"message"}, i.e. nothing
  will happen unless a signal handler is invoked (e.g. using
  \code{\link[base]{tryCatch}()} or
  \code{\link[base]{withCallingHandlers}()})


}
\section{Functions}{
\itemize{
\item \code{custom_stop}: raises a customized error and
\link[base:stop]{stops} execution of current expression

\item \code{custom_warning}: raises a customized warning and generates
the corresponding \link[base:warning]{warning message}

\item \code{custom_message}: raises a customized message and shows the 
corresponding \link[base:message]{diagnostic message}

\item \code{custom_signalCondition}: signals a customized condition. Note in this
case nothing will happen unless a custom signal handler was instantiated
(e.g. using \code{\link[base]{tryCatch}()} or
\code{\link[base]{withCallingHandlers}()})

\item \code{custom_exceptionHandler}: provides a generating function for customized
exception handlers
}}

\seealso{
\code{\link[base]{stop}()}, \code{\link[base]{warning}()}, or
  \code{\link[base]{message}()} for basic handling of errors, warnings, or
  messages, respectively
  
  \code{\link[base]{signalCondition}()} for signalling a condition
}
