% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendClass.R
\name{extend_class}
\alias{extend_class}
\alias{extend_class<-}
\title{Extend Classes}
\usage{
extend_class(subclass, class)

extend_class(x) <- value
}
\arguments{
\item{subclass, value}{character, \emph{unique} name(s) of subclass(es) that
extend the base classes \code{class}}

\item{class}{character, \emph{unique} names of classes to be extended}

\item{x}{any object, which's class is to be expanded}
}
\value{
a character string containing a extended list of classes
}
\description{
The function consistently extends classes. See details for algorithm.
}
\details{
Both \code{class} and \code{subclass} (or equivalently \code{value})
  must be an ordered set of subclasses. If \code{class} and \code{subclass}
  are \emph{not} disjoint, the intersection has to be of the same order.
  Note that if either \code{class} or \code{subclass} are not unique or the
  order of the intersection differs, an error is thrown.
  
  If (the first element of) \code{subclass} is contained in \code{class},
  the return value contain subclasses of \code{class} with \code{subclass}
  the first element, i.e. those elements of \code{class} that "come before"
  \code{subclass} omitted. Otherwise \code{subclass} will be prepended. For
  \code{subclass} of length greater 1, this will be done iteratively.
}
